requirements.txt
openai
openai-agents
chromadb
langchain-chroma
langchain
langchain-openai
pypdf
openai-agents[viz]
rich
mem0ai
unstructured[md] nltk
langchain-text-splitters